+++ src/ErrorOrX/ErrorOr.ToErrorOrExtensions.cs
@@ -15,14 +15,10 @@ public static partial class ErrorOrExtensions
     /// <summary>
     ///     Creates an <see cref="ErrorOr{TValue}" /> instance with the given <paramref name="errors" />.
     /// </summary>
-    /// <exception cref="ArgumentNullException">Thrown when <paramref name="errors" /> is null.</exception>
-    /// <exception cref="ArgumentException">Thrown when <paramref name="errors" /> is an empty list.</exception>
     public static ErrorOr<TValue> ToErrorOr<TValue>(this List<Error> errors) => errors;
 
     /// <summary>
     ///     Creates an <see cref="ErrorOr{TValue}" /> instance with the given <paramref name="errors" />.
     /// </summary>
-    /// <exception cref="ArgumentNullException">Thrown when <paramref name="errors" /> is null.</exception>
-    /// <exception cref="ArgumentException">Thrown when <paramref name="errors" /> is an empty array.</exception>
     public static ErrorOr<TValue> ToErrorOr<TValue>(this Error[] errors) => errors;
 }